[GreenPad-vc4] squashed merge of default branch GreenPad changes:
- [GreenPad] prevent negative positions when centering window (87ca17f)
- [GreenPad] always append log to prevent log being flushed when multi-opening (bbaa4cd)
- squashed merge of https://github.com/RamonUnch/GreenPad-vc4/pull/71 with custom add/removes, WIP. (9ec13f3)
- [GreenPad] remove OLE DnD handling inside main edit control (as I don't need this for now) (17fd3bd)
- [GreenPad] multiple changes: (0a10ce6)
- [GreenPad] free COM object itself properly when reference count turns to zero (b1d34b8)
- [GreenPad] partly sync with https://github.com/RamonUnch/GreenPad-vc4/pull/71/commits/03c40a87ff9b5b0d7d7134571f95e52e1cbe4ae1 and properly fill remaining bytes with zero. (6006a7c)
- [GreenPad] remove WM_RBUTTONDOWN handling in editwing, since I don't use OLE Drag here (375816e)
- [GreenPad] more preporcessor guards and sync with RamonUnch/GreenPad#71 (38d6ad8)
- [GreenPad] call `CoLockObjectExternal` in OleDnDTarget for win32s (5b4ba2b)
- [GreenPad] follow-up rev 9f006c63 (8e31f12)
- [GreenPad] fix non-OLE file drop (9eb80f8)
- [GreenPad] bring myself to foreground if same file is dropped to window (for OLE DnD enabled builds only) (0ad8510)
- [GreenPad] update VC71 project files (5a883ed)
- [GreenPad] OleDnDTarget: delete itself when refcnt goes to zero, add more loggings (888e5a1)
- [GreenPad] OleDnDTarget: fix logic in constructor, add more loggings (796fe39)
- [GreenPad] import "IDropTarget minor adjustmants" (a3a0b5a)
- [GreenPad] OLE drop target: add CF_HDROP to WM_DROPFILES redirection code (4e3878e)
- [GreenPad] OLE drop target: guard CF_HDROP handling with win32s detection as non-win32s windows converts OLE CF_HDROP to WM_DROPFILES for us. (d62706e)
- [GreenPad] vc71 project: update win32s OLE object and map paths (25b31ec)
- [GreenPad] add extra fields for identify myself after file list of DROPFILES, and revert changes in rev cbfc779bb and part of rev 312e238b (6a8420b)
- [GreenPad] don't need hwnd from view control, use hwnd from itself instead. (cb0e5fc)
- [GreenPad] log: win32s doesn't supprot FILE_APPEND_DATA, go back to use GENERIC_WRITE instead. (bf16f57)
- [GreenPad] change return type of DlgImpl::MainProc() to fix win64 build (2addd0e)
- [GreenPad] fix type when passing to on_drop() (3a296e4)
- [GreenPad] import "Make thunk read+exe only for security" (ee95477)
- [GreenPad] add casts to fix warnings (4d6c8aa)
- [GreenPad] follow-up rev 2addd0e3 (0a25c9f)
- [GreenPad] change cast type to match SetWindowPos() defination (1df6b23)

+++ GreenPad-vc4/OpenSaveDlg.cpp
@@ -582,7 +582,7 @@ void ReopenDlg::on_init()
 	for( ulong i=0; i<csl_.size(); ++i )
 		if( csl_[i].type & 1 ) // 2:=SAVE
 			cb.Add( csl_[i].longName );
-	if(csIndex_ < 0 || csIndex_ > csl_.size()) csIndex_ = 0;
+	if(csIndex_ < 0 || (ulong)csIndex_ > csl_.size()) csIndex_ = 0;
 	cb.Select( csl_[csIndex_].longName );
 }
